NICE-250 Refactor announcer code into a re-useable component
